Fix the update of nTodo / nTodoNested in the lists of resultsComponent class
Since the items of the lists in the resultsComponent instances
are references to objects of the root tree, their parent / list
refers to that of the root, and not to the resultsComponent
instance and lists. Hence, the latter do not get updated in terms
of nTodo and nTodoNested.

To fix this, these things have been done:

1) whenever the nTodo property of a todoListComponent instance
changes due to the increaseListNTodo and decreaseListNTodo methods
of todoComponent class, a this.list.getPubSubName("NTODO INCREASE",
"main", false) PubSub token is published, along with the item that
triggers the change. The lists of the results components are s
ubscribed to this token and when they capture it they change their
nTodo property accordingly if they contain the corresponding item.
This behaviour is similar to that used to update the results lists
when deleting elements.

2) when removing an item using the removeItem() from
baseListComponent (or inherited) class, distinguish the behaviour
whether the list is primary or not. A primary list is one in the
root tree. The lists in the results components are not primary.
A few operations are performed only if the list is primary (eg,
updating the parent dateOfEdit, notify the descendant in the dom
of their deletions, and so on).

3) the nTodoNested property is not updated in the lists of the
results components, since it does not make much sense. Indeed, the
list itself can store nested items, too.