Insert new items to displayed results lists if there is a match
We want to update the displayed results lists when new projects/
todo/nodes are added.

To do this, whenever a new item is inserted in a primary
baseListComponent (or inherited class) instance, a
this.getPubSubName("ADD ITEM", "main", false) PubSub token is
published, along with the item that triggers the change. Recall
that a list is 'primary' if it is part of the root tree, and it is
not a results list.

The lists of the results components are subscribed to this token
and when they capture it they insert the item in there if it
matches the corresponding filters.

This behaviour is similar to that used to update the results lists
when deleting elements.

The code related to this method in listInComponentMixin.js and
todoListComponent.js has been updated accordingly.