Insert/delete edited items in displayed results lists if there is a match
We want to update the displayed results lists when root items are
edited. Indeed, consider a filter of 'Low priority' todo items.
Editing the priority of an item should insert it to the displayed
filter results if its priority is set to Low, or delete it from it
if one of its items is no longer a 'Low priority' one.

A this.list.getPubSubName("EDIT ITEM", "main", false) PubSub token
is published whenever a title, description, tag, dueDate,
imminence, priority, state property of a baseComponent or inherited
class component changes, along with the item that triggers the
change.

The lists of the results components are subscribed to this token
and when they capture it:
  - if the item matches the corresponding filter and is not in the
    results list yet, it is inserted in there.
  - if the item does not match the corresponding filter but is in
    the results list, it is removed from there.

This behaviour is similar to that used to update the results lists
when deleting elements.