Add filtersAndTagsComponent and filtersAndTagsDomComponent modules
The filtersAndTagsComponent class is used to create data for a
'Filter and tags' view. It initializes the data to be used to
create a resultsComponent to, eg, filter all the todo with high
priority, or due today, or with a given tag, and so on.

The filtersAndTagsDomComponent class is used to create an
interface for such 'Filter and tags' view. It initializes a button
for each required filter and set a callback to initialize a
resultsComponent instance (using the data from the
filtersAndTagsComponent component) and display it.

An instance of these classes is temporarily created in the
homeDomComponent, just for testing, and will be removed later on.

No style is applied yet.