Add filtersAndTagsDomMiniNavComponent module
The filtersAndTagsDomMiniNavComponent class instantiate a div that
acts as a button which shows the filtersAndTagsDomComponent view
when clicking on it.

This behaviour is the same as the one obtained from the
projectDomMiniNavComponent class. In fact, it is added to it through
composition using the same domMiniComponentMixin module.

Some domMiniComponentMixin has been slightly modified. Mostly, the
this.initNTodoNestedIcon() div has been included only if the
this.initNTodoNestedIcon method is defined (it is not for the
filtersAndTagsDomComponent, as there is no todocount in there).

The homeDomComponent has been modified to show this mini-view div
instead of the whole Filters And Tags view.