Add resultsDomMiniNavComponent module
The resultsDomMiniNavComponent class instantiates a div that acts
as a button which shows the resultsDomComponent view when clicking
on it.

This behaviour is the same as the one obtained from the
filtersAndTagsDomComponent class. In fact, it is added to it
through composition using the same domMiniComponentMixin module.

domMiniComponentMixin has been again slightly modified. Mostly, the
this.initNTodoIcon() div has been included only if the
this.initNTodoNestedIcon method is not defined, but the
this.initNTodoIcon() is (it is for the resultsDomComponent, as
there is a todo count in there).

A temporary instance of this mini-view has been added to the
homeDomComponent instance, for testing.