Add searchComponent and searchDomComponent modules
The searchComponent module defines a class which can be used to
intantiate an object to lookup a string in the root component
(within the title, description, and tags), creating a proper
rootComponent object. An updateSearchResults(value) method is
used to update the lookup string (the results are indirectly
updated, too).

The searchDomComponent module defines a class which can be used
to instantiate an object to get the lookup string through an
input element, and display the search results. The results are
displayed in a resultsDomComponent view right now, but this will
be improved later on. Also, the displayed results change once the
lookup string changes.

Temporarily, such objects have been instantiated in the
homeDomComponent instance, for testing, and will be replaced in
the future.