Add overdue, today and upcoming views on homeDomComponent view
Each one is represented as a resultsDomMiniNavComponent div (to
have a link to the view showing just the corresponding results
list), and a resultsDomListComponent div (to quickly display the
corresponding items directly in the homePage, since they are
important info to show the user).

Minor style changes have been applied.