Fix searchComponent and searchDomComponent
The lists in the searchDomComponent view were not properly updating
when, eg, deleting them or changing their state. This was due to
the associated PubSub tokens being subsribed when creating the
home view, but deleted when moving in the search view (as doing
that unsubscribed everything from the 'main:' topic).

This has been fixed by moving the creation of the resultsComponent
instance within the searchDomComponent instance, similarly to what
is done in the filtersAndTagsDomComponent, although this introduces
some coupling.