Add possibility to sort the list items according to a given criteria
First, a getSortBy() method has been added to the baseListComponent
class to retrieve a new array of the items ordered according to a
given criteria, eg, by 'title' or by 'due date' (if it is a
todoListComponent), and optionally in descending order. The
ordering callback is hard-coded in a static object of the class
sortCallbacks, and new specific callbacks are added for the
todoListComponent class.

Then, a 'sortBy' method has been added to the baseListDomComponent
class to reorder its displayed items. This will be achieved by
using the ordered array obtained from the above mentioned method
of the associated baseListComponent instance, and properly setting
the 'order' property of the corresponding <li> element containing
the items representation.

The reordering is triggered in two occasions: when adding a new
item to the list, or when modifying it. Note that removing an item
does not really change the order of the other elements.

As of now, however, it has been temporarily imposed that the
ordering is 'by Title, in ascending order'. A user selection UI
will be added next.