Redefine initOtherInfo method for mini dom components through domMiniComponentMixin mixin
The size of each list of the instances created with a class
extended via composition through this mixin is displayed in the
'other info' section in the dom component.

As a result, for the classes using this mixin (eg,
projectDomMiniComponent, todoDomMiniComponent,...), the number of
sub/projects, sub-tasks, or note replies, if any, is displayed.

If this number is 0, it gets hidden: this is achieved through css.